requirements.txt
requests
python-dotenv
fastapi
openpyxl
pandas
python-multipart
streamlit
uvicorn